Finish cherry-picking Asserting/Assuming changes from Bob's branch
